Added support for tweet importing
