+++ components/common/CharmEditor/components/tweet/tweetSpec.ts
@@ -8,21 +8,6 @@ export type TweetNodeAttrs = {
   screenName: string;
 };
 
-// a function to extract user screen name and tweet id from a tweet url
-export function extractTweetAttrs(url: string): TweetNodeAttrs | null {
-  if (!url) {
-    return null;
-  }
-
-  const match = url.match(/twitter\.com\/([^/]+)\/status\/(\d+)/);
-  if (!match) {
-    return null;
-  }
-  return {
-    screenName: match[1],
-    id: match[2]
-  };
-}
 export function spec(): RawSpecs {
   return {
     type: 'node',
